requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Bcrypt==1.0.1
Flask-SocketIO==5.3.4
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
SQLAlchemy==2.0.20
twilio==8.5.0
openai==0.28.1
pandas==2.0.3
numpy==1.24.3
pydub==0.25.1
gunicorn==21.2.0
celery==5.3.4
redis==4.6.0
Pillow==10.0.0
requests==2.31.0
pytest==7.4.0
python-magic==0.4.27
email-validator==2.0.0.post2
xlrd==2.0.1
openpyxl==3.1.2
whisper==1.1.10
